% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blobs.R
\name{blob_delete}
\alias{blob_delete}
\title{Delete a blob}
\usage{
blob_delete(storage_account, container_name, azure_access_key, path)
}
\arguments{
\item{storage_account}{The name of the storage account}

\item{container_name}{The name of the container}

\item{azure_access_key}{Access key to the storage account}

\item{path}{Path to the blob}
}
\value{
The result of the request is returned invisibly.
}
\description{
Downloads the blob speficied in \code{resource_path}.
}
