% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blobs.R
\name{blob_download}
\alias{blob_download}
\title{Download a blob}
\usage{
blob_download(
  storage_account,
  container_name,
  azure_access_key,
  resource_path,
  dest_file = character(0)
)
}
\arguments{
\item{storage_account}{The name of the storage account}

\item{container_name}{The name of the container}

\item{azure_access_key}{Access key to the storage account}

\item{resource_path}{Path to the blob}

\item{dest_file}{The name of the destination file. If empty, extracts the
file name from \code{resource_path}.}
}
\value{
The result of the request is returned invisibly.
}
\description{
Downloads the blob speficied in \code{resource_path}.
}
