% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blobs.R
\name{blob_parse}
\alias{blob_parse}
\title{Parse blob metadata}
\usage{
blob_parse(blob)
}
\arguments{
\item{blob}{A list containing the blob metadata. (If there are no blobs,
the result is a tibble with 0 rows and 0 columns).}
}
\description{
Extracts blob metadata from a list obtained from parsing the xml result from
the blob service API. Used as a helper function for \code{\link[=blobs_list]{blobs_list()}}.
}
