% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blobs.R
\name{blob_upload}
\alias{blob_upload}
\title{Upload a blob}
\usage{
blob_upload(
  storage_account,
  container_name,
  azure_access_key,
  source_file_path,
  destination_path = "/"
)
}
\arguments{
\item{storage_account}{The name of the storage account}

\item{container_name}{The name of the container}

\item{azure_access_key}{Access key to the storage account}

\item{source_file_path}{Path to the file to be uploaded}

\item{destination_path}{Destination path for the blob}
}
\value{
The result of the request is returned invisibly.
}
\description{
Uploads the blob speficied in \code{source_file_path} to the \code{destination_path}.
The file name is not changed.
}
