% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blobs.R
\name{blobs_list}
\alias{blobs_list}
\title{List blobs}
\usage{
blobs_list(storage_account, container_name, azure_access_key)
}
\arguments{
\item{storage_account}{The name of the storage account}

\item{container_name}{The name of the container}

\item{azure_access_key}{Access key to the storage account}
}
\value{
A tibble with information of the blobs.
}
\description{
Lists all the blobs in the given container. Ignores blobs with content length
\code{0}.
}
