% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blob_helpers.R
\name{calc_signature}
\alias{calc_signature}
\title{Calculate the signature in the request}
\usage{
calc_signature(string_to_sign, azure_access_key)
}
\arguments{
\item{string_to_sign}{Obtained with \code{\link[=get_string_to_sign]{get_string_to_sign()}}}

\item{azure_access_key}{Access key to the storage account}
}
\description{
See \href{https://docs.microsoft.com/en-us/rest/api/storageservices/authorize-with-shared-key#specifying-the-authorization-header}{Specifying the authorization header.}
in the documentation.
}
