% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blob_helpers.R
\name{construct_url}
\alias{construct_url}
\title{Construct the url in the correct format}
\usage{
construct_url(storage_account, container_name, rest = "")
}
\arguments{
\item{storage_account}{The name of the storage account}

\item{container_name}{The name of the container}

\item{rest}{Rest of the resource. If referring to a file, the destination file path. If listing files, the required parameters.}
}
\description{
Constructs the url. For listing blobs adds the desired parameters and others the file path.
}
