% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blob_helpers.R
\name{get_canonicalized_headers}
\alias{get_canonicalized_headers}
\title{Canonicalized headers}
\usage{
get_canonicalized_headers(blob_type, date, version)
}
\arguments{
\item{blob_type}{Always \code{"BlockBob"}, which corresponds to a new block
(as apposed to page blob or append blob)}

\item{date}{Current timestamp formatted with \code{"\%a, \%d \%b \%Y \%H:\%M:\%S \%Z"}}

\item{version}{Always \code{"2020-04-08"}, most recent version when the package
was initially implemented}
}
\description{
\href{https://docs.microsoft.com/en-us/rest/api/storageservices/authorize-with-shared-key}{As speficied in the documentation.}.
}
