% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blob_helpers.R
\name{get_canonicalized_resource}
\alias{get_canonicalized_resource}
\title{Canonicalized resource}
\usage{
get_canonicalized_resource(storage_account, container_name, rest = "")
}
\arguments{
\item{storage_account}{The name of the storage account}

\item{container_name}{The name of the container}

\item{rest}{Rest of the resource. If referring to a file, the destination file path. If listing files, the required parameters.}
}
\description{
\href{https://docs.microsoft.com/en-us/rest/api/storageservices/authorize-with-shared-key}{As speficied in the documentation.}.
}
