% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blob_helpers.R
\name{get_string_to_sign}
\alias{get_string_to_sign}
\title{Get the string to be signed}
\usage{
get_string_to_sign(
  canonicalized_headers,
  canonicalized_resource,
  http_method = "GET",
  content_length = "",
  content_md5 = "",
  content_type = ""
)
}
\arguments{
\item{canonicalized_headers}{Obtained with \code{\link[=get_canonicalized_headers]{get_canonicalized_headers()}}}

\item{canonicalized_resource}{Obtained with \code{\link[=get_canonicalized_resource]{get_canonicalized_resource()}}}

\item{http_method}{\code{"GET"}, \code{"PUT"}, \code{"DELETE"} etc.}

\item{content_length}{Length of the file to be uploaded in bytes. Otherwise (and by default) empty}

\item{content_md5}{md5sum of the file to be uploaded. Otherwise (and by default) empty}

\item{content_type}{Content type of the file to be uploaded. Otherwise (and by default) empty}
}
\description{
See \href{https://docs.microsoft.com/en-us/rest/api/storageservices/authorize-with-shared-key#constructing-the-signature-string}{Constructing the signature string}
in the documentation.
}
